S02P31B105-106 | 200610 | 에러 수정

See merge request s02-final/s02p31b105!164

+++ Backend/wiselife/src/main/java/com/ssafy/wiselife/service/ReviewServiceImpl.java
@@ -110,7 +110,7 @@ public class ReviewServiceImpl implements IReviewService {
 			reviewrepo.save(reviewEntity);
 
 			// 별점 평균 저장
-			int avgScore = reviewrepo.avgScoreMeeting(review.getMeetingId());
+			double avgScore = reviewrepo.avgScoreMeeting(review.getMeetingId());
 			meeting.setScore(avgScore);
 			meetingrepo.save(meeting);
